+++ pose_estimation/trainers/assembler.py
@@ -18,8 +18,8 @@
 from ..model import PEModel, PETrainer
 from ..generators.pose_estimation import RIterator
 from pose_estimation.model import BinaryHeatmapLayer, GaussHeatmapLayer, V2PAFLayer
-from makiflow.core import MakiRestorable, TrainerBuilder
-from makiflow.core import MakiTensor, MakiBuilder
+from pose_estimation.model.postprocess_modules import TFPostProcessModule
+from makiflow.core import MakiRestorable, TrainerBuilder, MakiTensor
 from makiflow.layers import InputLayer
 from makiflow.distillation.core import DistillatorBuilder
 
@@ -92,15 +92,21 @@ class ModelAssembler:
         shape = gen_layer.get_shape()
         # Change batch_size to 1
         shape[0] = 1
-        # Change image size to dymanic size
+        # Change image size to dynamic size
         shape[1] = None
         shape[2] = None
         name = gen_layer.get_name()
 
         input_layer = InputLayer(input_shape=shape, name=name)
-
-        model = PEModel.from_json(model_config[ModelAssembler.ARCH_PATH], input_layer)
+        tf_post_process = TFPostProcessModule()
+        model = PEModel.from_json(
+            model_config[ModelAssembler.ARCH_PATH],
+            postprocess_class=tf_post_process,
+            input_tensor=input_layer
+        )
         model.set_session(sess)
+        tf_post_process.set_session(sess)
+        tf_post_process.set_paf_heatmap(model.get_main_paf_tensor(), model.get_main_heatmap_tensor())
 
         # Load pretrained weights
         weights_path = model_config[ModelAssembler.WEIGHTS_PATH]
